Refactor resources
